pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-aiplatform>=1.120.0",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-storage>=2.19.0",
    "python-dotenv",
    "pandas",
    "datasets>=4.2.0",
    "gcsfs>=2025.9.0",
    "dotenv>=0.9.9",
    "kfp>=2.14.6",
]
